outguess: update livecheck
The existing `livecheck` block for `outguess` checks the homepage but
requests to www.rbcafe.com fail with the default user agent, so this
check fails and retries with the `:browser` user agent (which works).
This adds `user_agent: :browser` to the `livecheck` block URL, so it
will continue to work after we remove the user agent fallback behavior
in livecheck.